+++ core/src/main/java/tech/pegasys/web3signer/core/service/jsonrpc/handlers/HeaderHelpers.java
@@ -16,12 +16,11 @@ import java.util.List;
 
 import com.google.common.net.HttpHeaders;
 import io.vertx.core.MultiMap;
-import io.vertx.core.http.impl.headers.HeadersMultiMap;
 
 public class HeaderHelpers {
 
   public static MultiMap createHeaders(final MultiMap headers) {
-    final MultiMap headersToReturn = new HeadersMultiMap();
+    final MultiMap headersToReturn = MultiMap.caseInsensitiveMultiMap();
     headers.forEach(entry -> headersToReturn.add(entry.getKey(), entry.getValue()));
 
     headersToReturn.remove(HttpHeaders.CONTENT_LENGTH);