+++ signing/src/test/java/tech/pegasys/web3signer/signing/config/metadata/parser/YamlSignerParserTest.java
@@ -242,7 +242,8 @@ class YamlSignerParserTest {
 
     Assertions.assertThatThrownBy(
             () -> signerParser.parse(signerParser.readSigningMetadata(yamlMetadata)))
-        .hasRootCauseMessage("Provider \"https\" not installed");
+        .rootCause()
+        .hasMessageContaining("scheme 'https' not allowed for Path deserialization");
   }
 
   @Test